+++ lib/bottom_sheet.dart
@@ -12,8 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-library bottom_sheet;
-
-export 'package:bottom_sheet/src/flexible_bottom_sheet.dart';
-export 'package:bottom_sheet/src/flexible_bottom_sheet_header_delegate.dart';
-export 'package:bottom_sheet/src/flexible_bottom_sheet_route.dart';
+export 'package:bottom_sheet/src/flexible_bottom_sheet.dart'
+    show
+        FlexibleBottomSheet,
+        FlexibleDraggableScrollableHeaderWidgetBuilder,
+        FlexibleDraggableScrollableWidgetBodyBuilder,
+        FlexibleDraggableScrollableWidgetBuilder;
+export 'package:bottom_sheet/src/flexible_bottom_sheet_header_delegate.dart'
+    show FlexibleBottomSheetHeaderDelegate;
+export 'package:bottom_sheet/src/flexible_bottom_sheet_route.dart'
+    show showFlexibleBottomSheet, showStickyFlexibleBottomSheet;