+++ lib/src/flexible_bottom_sheet.dart
@@ -89,7 +89,7 @@ typedef FlexibleDraggableScrollableWidgetBodyBuilder = SliverChildDelegate
 /// [bottomSheetColor] - bottom sheet color. If you want to make rounded edges,
 /// pass a [Colors.transparent] here, and set the color and border radius of
 /// the bottom sheet in the [decoration].
-class FlexibleBottomSheet extends StatefulWidget {
+class FlexibleBottomSheet<T> extends StatefulWidget {
   final double minHeight;
   final double initHeight;
   final double maxHeight;
@@ -107,9 +107,13 @@ class FlexibleBottomSheet extends StatefulWidget {
   final VoidCallback? onDismiss;
   final Color? keyboardBarrierColor;
   final Color? bottomSheetColor;
+  final PopupRoute<T>? route;
+  final bool useRootScaffold;
+  final BorderRadiusGeometry? bottomSheetBorderRadius;
 
   FlexibleBottomSheet({
-    Key? key,
+    super.key,
+    this.route,
     this.minHeight = 0,
     this.initHeight = 0.5,
     this.maxHeight = 1,
@@ -126,17 +130,19 @@ class FlexibleBottomSheet extends StatefulWidget {
     this.onDismiss,
     this.keyboardBarrierColor,
     this.bottomSheetColor,
+    this.bottomSheetBorderRadius,
     this.draggableScrollableController,
+    this.useRootScaffold = true,
   })  : assert(minHeight >= 0 && minHeight <= 1),
         assert(maxHeight > 0 && maxHeight <= 1),
         assert(maxHeight > minHeight),
         assert(!isCollapsible || minHeight == 0),
         assert(anchors == null || !anchors.any((anchor) => anchor > maxHeight)),
         assert(anchors == null || !anchors.any((anchor) => anchor < minHeight)),
-        assert(isExpand == true || maxHeight == initHeight && anchors == null),
-        super(key: key);
+        assert(isExpand || maxHeight == initHeight && anchors == null);
 
   FlexibleBottomSheet.collapsible({
+    required PopupRoute<T> route,
     Key? key,
     double initHeight = 0.5,
     double maxHeight = 1,
@@ -152,7 +158,10 @@ class FlexibleBottomSheet extends StatefulWidget {
     Decoration? decoration,
     Color? keyboardBarrierColor,
     Color? bottomSheetColor,
+    bool useRootScaffold = true,
+    BorderRadiusGeometry? bottomSheetBorderRadius,
   }) : this(
+          route: route,
           key: key,
           maxHeight: maxHeight,
           draggableScrollableController: draggableScrollableController,
@@ -170,13 +179,15 @@ class FlexibleBottomSheet extends StatefulWidget {
           decoration: decoration,
           keyboardBarrierColor: keyboardBarrierColor,
           bottomSheetColor: bottomSheetColor,
+          useRootScaffold: useRootScaffold,
+          bottomSheetBorderRadius: bottomSheetBorderRadius,
         );
 
   @override
-  _FlexibleBottomSheetState createState() => _FlexibleBottomSheetState();
+  State<FlexibleBottomSheet<T>> createState() => _FlexibleBottomSheetState();
 }
 
-class _FlexibleBottomSheetState extends State<FlexibleBottomSheet> {
+class _FlexibleBottomSheetState<T> extends State<FlexibleBottomSheet<T>> {
   late final DraggableScrollableController _controller;
 
   late final WidgetsBinding _widgetBinding;
@@ -195,84 +206,11 @@ class _FlexibleBottomSheetState extends State<FlexibleBottomSheet> {
     widget.animationController?.addStatusListener(_animationStatusListener);
   }
 
-  @override
-  Widget build(BuildContext context) {
-    return NotificationListener<DraggableScrollableNotification>(
-      onNotification: _scrolling,
-      child: DraggableScrollableSheet(
-        maxChildSize: _currentMaxChildSize,
-        minChildSize: widget.minHeight,
-        initialChildSize: _initialChildSize,
-        snap: widget.anchors != null,
-        controller: _controller,
-        snapSizes: widget.anchors,
-        expand: widget.isExpand,
-        builder: (
-          context,
-          controller,
-        ) {
-          return ChangeInsetsDetector(
-            handler: (change) {
-              final inset = change.currentInset;
-              final delta = change.delta;
-
-              if (delta > 0 && !_isClosing) {
-                _animateToMaxHeight();
-                _widgetBinding.addPostFrameCallback(
-                  (_) {
-                    _animateToFocused(controller);
-                  },
-                );
-              }
-              // checking for openness of the keyboard before opening the sheet
-              if (delta == 0 && inset > 0) {
-                _widgetBinding.addPostFrameCallback(
-                  (_) {
-                    setState(
-                      () {
-                        _initialChildSize = widget.maxHeight;
-                      },
-                    );
-                  },
-                );
-              }
-            },
-            child: Scaffold(
-              backgroundColor: widget.bottomSheetColor ??
-                  Theme.of(context).bottomSheetTheme.backgroundColor ??
-                  Theme.of(context).backgroundColor,
-              body: _Content(
-                builder: widget.builder,
-                decoration: widget.decoration,
-                bodyBuilder: widget.bodyBuilder,
-                headerBuilder: widget.headerBuilder,
-                minHeaderHeight: widget.minHeaderHeight,
-                maxHeaderHeight: widget.maxHeaderHeight,
-                currentExtent: _controller.isAttached
-                    ? _controller.size
-                    : widget.initHeight,
-                scrollController: controller,
-                cacheExtent: _calculateCacheExtent(
-                  MediaQuery.of(context).viewInsets.bottom,
-                ),
-                getContentHeight:
-                    !widget.isExpand ? _changeInitAndMaxHeight : null,
-              ),
-            ),
-          );
-        },
-      ),
-    );
-  }
-
-  @override
-  void dispose() {
-    widget.animationController?.removeStatusListener(_animationStatusListener);
-    super.dispose();
-  }
-
   // Method will be called when scrolling.
   bool _scrolling(DraggableScrollableNotification notification) {
+    /// Force widget rebuild to change bottomSheetOffset value
+    setState(() {});
+
     if (_isClosing) return false;
 
     _initialChildSize = notification.extent;
@@ -304,10 +242,10 @@ class _FlexibleBottomSheetState extends State<FlexibleBottomSheet> {
     if (FocusManager.instance.primaryFocus == null || _isClosing) return;
 
     _widgetBinding.addPostFrameCallback((_) {
-      final keyboardHeight = MediaQuery.of(context).viewInsets.bottom;
+      final keyboardHeight = MediaQuery.viewInsetsOf(context).bottom;
       final widgetHeight = FocusManager.instance.primaryFocus!.size.height;
       final widgetOffset = FocusManager.instance.primaryFocus!.offset.dy;
-      final screenHeight = MediaQuery.of(context).size.height;
+      final screenHeight = MediaQuery.sizeOf(context).height;
 
       final targetWidgetOffset =
           screenHeight - keyboardHeight - widgetHeight - 20;
@@ -344,14 +282,16 @@ class _FlexibleBottomSheetState extends State<FlexibleBottomSheet> {
 
   void _dismiss() {
     if (widget.isCollapsible) {
-      if (widget.onDismiss != null) widget.onDismiss!();
+      widget.onDismiss?.call();
       Navigator.maybePop(context);
+      WidgetsBinding.instance
+          .addPostFrameCallback((_) => widget.route?.changedInternalState());
     }
   }
 
   void _changeInitAndMaxHeight(double height) {
     if (!widget.isExpand) {
-      final screenHeight = MediaQuery.of(context).size.height;
+      final screenHeight = MediaQuery.sizeOf(context).height;
 
       final fractionalValue = height / screenHeight;
       if (fractionalValue < _currentMaxChildSize) {
@@ -372,6 +312,130 @@ class _FlexibleBottomSheetState extends State<FlexibleBottomSheet> {
     } else {
       return defaultExtent;
     }
+  }
+
+  @override
+  void dispose() {
+    widget.animationController?.removeStatusListener(_animationStatusListener);
+    _controller.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final bottomSheetThemeBackground =
+        Theme.of(context).bottomSheetTheme.backgroundColor;
+    final colorSchemeBackground = Theme.of(context).colorScheme.background;
+
+    final bottomSheetColor = widget.bottomSheetColor ??
+        bottomSheetThemeBackground ??
+        colorSchemeBackground;
+    final contentDecoration = widget.decoration ??
+        BoxDecoration(
+          color: widget.bottomSheetColor ??
+              bottomSheetThemeBackground ??
+              colorSchemeBackground,
+        );
+
+    return NotificationListener<DraggableScrollableNotification>(
+      onNotification: _scrolling,
+      child: DraggableScrollableSheet(
+        maxChildSize: _currentMaxChildSize,
+        minChildSize: widget.minHeight,
+        initialChildSize: _initialChildSize,
+        snap: widget.anchors != null,
+        controller: _controller,
+        snapSizes: widget.anchors,
+        expand: widget.isExpand,
+        builder: (
+          context,
+          controller,
+        ) {
+          return ChangeInsetsDetector(
+            handler: (change) {
+              final inset = change.currentInset;
+              final delta = change.delta;
+
+              if (delta > 0 && !_isClosing) {
+                _animateToMaxHeight();
+                _widgetBinding.addPostFrameCallback(
+                  (_) {
+                    _animateToFocused(controller);
+                  },
+                );
+              }
+              // Checking for openness of the keyboard before opening the sheet.
+              if (delta == 0 && inset > 0) {
+                _widgetBinding.addPostFrameCallback(
+                  (_) {
+                    setState(
+                      () {
+                        _initialChildSize = widget.maxHeight;
+                      },
+                    );
+                  },
+                );
+              }
+            },
+            child: Material(
+              type: MaterialType.transparency,
+              color: bottomSheetColor,
+              borderRadius: widget.bottomSheetBorderRadius,
+              clipBehavior: widget.bottomSheetBorderRadius != null
+                  ? Clip.antiAlias
+                  : Clip.none,
+              child: _RegisterScaffold(
+                useRootScaffold: widget.useRootScaffold,
+                backgroundColor: bottomSheetColor,
+                child: _Content(
+                  builder: widget.builder,
+                  decoration: contentDecoration,
+                  bodyBuilder: widget.bodyBuilder,
+                  headerBuilder: widget.headerBuilder,
+                  minHeaderHeight: widget.minHeaderHeight,
+                  maxHeaderHeight: widget.maxHeaderHeight,
+                  currentExtent: _controller.isAttached
+                      ? _controller.size
+                      : widget.initHeight,
+                  scrollController: controller,
+                  cacheExtent: _calculateCacheExtent(
+                    MediaQuery.viewInsetsOf(context).bottom,
+                  ),
+                  getContentHeight:
+                      !widget.isExpand ? _changeInitAndMaxHeight : null,
+                ),
+              ),
+            ),
+          );
+        },
+      ),
+    );
+  }
+}
+
+/// Register [Scaffold] for [FlexibleBottomSheet].
+class _RegisterScaffold extends StatelessWidget {
+  final bool useRootScaffold;
+  final Widget child;
+  final Color backgroundColor;
+
+  const _RegisterScaffold({
+    required this.useRootScaffold,
+    required this.child,
+    required this.backgroundColor,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return useRootScaffold
+        ? Scaffold(
+            backgroundColor: backgroundColor,
+            body: child,
+          )
+        : ColoredBox(
+            color: backgroundColor,
+            child: child,
+          );
   }
 }
 
@@ -399,8 +463,7 @@ class _Content extends StatefulWidget {
     this.minHeaderHeight,
     this.maxHeaderHeight,
     this.getContentHeight,
-    Key? key,
-  }) : super(key: key);
+  });
 
   @override
   State<_Content> createState() => _ContentState();
@@ -416,7 +479,7 @@ class _ContentState extends State<_Content> {
       WidgetsBinding.instance.addPostFrameCallback(
         (timeStamp) {
           final renderContent =
-              _contentKey.currentContext!.findRenderObject() as RenderBox;
+              _contentKey.currentContext!.findRenderObject()! as RenderBox;
           widget.getContentHeight!(renderContent.size.height);
         },
       );