+++ lib/src/flexible_bottom_sheet_route.dart
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+// ignore_for_file: format-comment
 import 'package:bottom_sheet/src/flexible_bottom_sheet.dart';
 import 'package:flutter/material.dart';
 
@@ -36,11 +37,13 @@ const Duration _bottomSheetDuration = Duration(milliseconds: 500);
 /// [isModal] - if true, overlay background with dark color.
 /// [anchors] - list of sizes in fractional value that the bottom sheet can accept.
 /// [keyboardBarrierColor] - keyboard color.
+/// [bottomSheetBorderRadius] - bottom sheet border radius.
 /// [bottomSheetColor] - bottom sheet color.
 /// [barrierColor] - barrier color.
 /// [duration] - animation speed when opening bottom sheet.
 /// [isSafeArea] - should the bottom sheet provide a SafeArea, false by default.
-/// [decoration] - BottomSheet decoration.
+/// [decoration] - content decoration bottom sheet.
+/// [useRootScaffold] - if true, add Scaffold widget on widget tree. Default true.
 Future<T?> showFlexibleBottomSheet<T>({
   required BuildContext context,
   required FlexibleDraggableScrollableWidgetBuilder builder,
@@ -56,10 +59,12 @@ Future<T?> showFlexibleBottomSheet<T>({
   List<double>? anchors,
   Color? keyboardBarrierColor,
   Color? bottomSheetColor,
+  BorderRadiusGeometry? bottomSheetBorderRadius,
   Color? barrierColor,
   Duration? duration,
   bool isSafeArea = false,
   BoxDecoration? decoration,
+  bool useRootScaffold = true,
 }) {
   assert(debugCheckHasMediaQuery(context));
   assert(debugCheckHasMaterialLocalizations(context));
@@ -81,10 +86,12 @@ Future<T?> showFlexibleBottomSheet<T>({
       anchors: anchors,
       keyboardBarrierColor: keyboardBarrierColor,
       bottomSheetColor: bottomSheetColor,
+      bottomSheetBorderRadius: bottomSheetBorderRadius,
       barrierBottomSheetColor: barrierColor,
       duration: duration,
       isSafeArea: isSafeArea,
       decoration: decoration,
+      useRootScaffold: useRootScaffold,
     ),
   );
 }
@@ -107,17 +114,19 @@ Future<T?> showFlexibleBottomSheet<T>({
 /// in which case the bottom sheet will calculate its height based on the content,
 /// but no more than [maxHeight] and [initHeight].
 /// [anchors] - list of sizes in fractional value that the bottom sheet can accept.
-/// [decoration] - BottomSheet decoration.
+/// [decoration] - content decoration bottom sheet.
 /// [minHeaderHeight] - minimum head size.
 /// [maxHeaderHeight] - maximum head size.
 /// [headerHeight] - head size.
 /// Set both [minHeaderHeight] and [maxHeaderHeight].
 /// Set one ([maxHeaderHeight] or [headerHeight]).
 /// [keyboardBarrierColor] - keyboard color.
+/// [bottomSheetBorderRadius] - bottom sheet border radius.
 /// [bottomSheetColor] - bottom sheet color.
 /// [barrierColor] - barrier color, if you pass [barrierColor] - [isModal] must be true.
 /// [duration] - animation speed when opening bottom sheet.
 /// [isSafeArea] - should the bottom sheet provide a SafeArea, false by default.
+/// [useRootScaffold] - if true, add Scaffold widget on widget tree. Default true.
 Future<T?> showStickyFlexibleBottomSheet<T>({
   required BuildContext context,
   required FlexibleDraggableScrollableHeaderWidgetBuilder headerBuilder,
@@ -138,9 +147,11 @@ Future<T?> showStickyFlexibleBottomSheet<T>({
   Decoration? decoration,
   Color? keyboardBarrierColor,
   Color? bottomSheetColor,
+  BorderRadiusGeometry? bottomSheetBorderRadius,
   Color? barrierColor,
   Duration? duration,
   bool isSafeArea = false,
+  bool useRootScaffold = true,
 }) {
   assert(maxHeaderHeight != null || headerHeight != null);
   assert(debugCheckHasMediaQuery(context));
@@ -167,9 +178,11 @@ Future<T?> showStickyFlexibleBottomSheet<T>({
       decoration: decoration,
       keyboardBarrierColor: keyboardBarrierColor,
       bottomSheetColor: bottomSheetColor,
+      bottomSheetBorderRadius: bottomSheetBorderRadius,
       barrierBottomSheetColor: barrierColor,
       duration: duration,
       isSafeArea: isSafeArea,
+      useRootScaffold: useRootScaffold,
     ),
   );
 }
@@ -194,9 +207,11 @@ class _FlexibleBottomSheetRoute<T> extends PopupRoute<T> {
   final ThemeData? theme;
   final Color? keyboardBarrierColor;
   final Color? bottomSheetColor;
+  final BorderRadiusGeometry? bottomSheetBorderRadius;
   final Color? barrierBottomSheetColor;
   final Duration? duration;
   final bool isSafeArea;
+  final bool useRootScaffold;
 
   @override
   final String? barrierLabel;
@@ -223,6 +238,7 @@ class _FlexibleBottomSheetRoute<T> extends PopupRoute<T> {
     required this.isExpand,
     required this.isModal,
     required this.isSafeArea,
+    required this.useRootScaffold,
     this.draggableScrollableController,
     this.builder,
     this.headerBuilder,
@@ -235,17 +251,18 @@ class _FlexibleBottomSheetRoute<T> extends PopupRoute<T> {
     this.decoration,
     this.keyboardBarrierColor,
     this.bottomSheetColor,
+    this.bottomSheetBorderRadius,
     this.barrierBottomSheetColor,
     this.duration,
-    RouteSettings? settings,
-  }) : super(settings: settings);
+    super.settings,
+  });
 
   @override
   AnimationController createAnimationController() {
     _animationController = AnimationController(
       duration: transitionDuration,
       debugLabel: 'FlexibleBottomSheet',
-      vsync: navigator?.overlay as TickerProvider,
+      vsync: navigator!.overlay!,
     );
 
     return _animationController;
@@ -262,6 +279,7 @@ class _FlexibleBottomSheetRoute<T> extends PopupRoute<T> {
       removeTop: true,
       child: isCollapsible
           ? FlexibleBottomSheet.collapsible(
+              route: this,
               initHeight: initHeight,
               maxHeight: maxHeight,
               builder: builder,
@@ -276,8 +294,11 @@ class _FlexibleBottomSheetRoute<T> extends PopupRoute<T> {
               decoration: decoration,
               keyboardBarrierColor: keyboardBarrierColor,
               bottomSheetColor: bottomSheetColor,
+              useRootScaffold: useRootScaffold,
+              bottomSheetBorderRadius: bottomSheetBorderRadius,
             )
           : FlexibleBottomSheet(
+              route: this,
               minHeight: minHeight,
               initHeight: initHeight,
               maxHeight: maxHeight,
@@ -293,6 +314,8 @@ class _FlexibleBottomSheetRoute<T> extends PopupRoute<T> {
               decoration: decoration,
               keyboardBarrierColor: keyboardBarrierColor,
               bottomSheetColor: bottomSheetColor,
+              useRootScaffold: useRootScaffold,
+              bottomSheetBorderRadius: bottomSheetBorderRadius,
             ),
     );
 
@@ -312,7 +335,7 @@ class _FlexibleBottomSheetRoute<T> extends PopupRoute<T> {
     Animation<double> secondaryAnimation,
     Widget child,
   ) {
-    const begin = Offset(0.0, 1.0);
+    const begin = Offset(0, 1);
     const end = Offset.zero;
     const curve = Curves.ease;
     final tween = Tween(begin: begin, end: end).chain(CurveTween(curve: curve));