+++ test/bottom_sheet_test.dart
@@ -1,6 +1,6 @@
 // Copyright (c) 2019-present,  SurfStudio LLC
 //
-// Licensed under the Apache License, Version 2.0 (the "License");
+// Licensed under the Apache License, Version 2 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
@@ -11,8 +11,6 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-
-import 'dart:ui';
 
 import 'package:bottom_sheet/bottom_sheet.dart';
 import 'package:flutter/material.dart';
@@ -35,6 +33,13 @@ void main() {
     ),
   );
 
+  void showSnackBar() {
+    ScaffoldMessenger.of(savedContext).showSnackBar(const SnackBar(
+      content: Text('SnackBar'),
+      duration: Duration(seconds: 5),
+    ));
+  }
+
   Future<void> showBottomSheet({
     bool? isCollapsible,
     bool? isDismissible,
@@ -45,8 +50,10 @@ void main() {
     bool? isModal,
     Color? barrierColor,
     List<double>? anchors,
+    bool? useRootScaffold,
   }) {
     return showFlexibleBottomSheet<void>(
+      useRootScaffold: useRootScaffold ?? true,
       minHeight: minHeight ?? 0,
       initHeight: initHeight ?? 0.5,
       maxHeight: maxHeight ?? 0.8,
@@ -82,11 +89,11 @@ void main() {
         (tester) async {
           await tester.pumpWidget(
             makeTestableWidget(
-              FlexibleBottomSheet(),
+              FlexibleBottomSheet<void>(),
             ),
           );
 
-          expect(() => FlexibleBottomSheet, returnsNormally);
+          expect(() => FlexibleBottomSheet<void>, returnsNormally);
 
           final flexibleScrollNotifier = find
               .byType(NotificationListener<DraggableScrollableNotification>);
@@ -104,11 +111,11 @@ void main() {
         unawaited(showBottomSheet());
 
         await tester.pumpAndSettle();
-        expect(find.byType(FlexibleBottomSheet), findsOneWidget);
+        expect(find.byType(FlexibleBottomSheet<void>), findsOneWidget);
 
-        await tester.tap(find.byType(FlexibleBottomSheet));
+        await tester.tap(find.byType(FlexibleBottomSheet<void>));
         await tester.pumpAndSettle();
-        expect(find.byType(FlexibleBottomSheet), findsOneWidget);
+        expect(find.byType(FlexibleBottomSheet<void>), findsOneWidget);
       });
 
       testWidgets(
@@ -121,12 +128,12 @@ void main() {
           ));
 
           await tester.pumpAndSettle();
-          expect(find.byType(FlexibleBottomSheet), findsOneWidget);
+          expect(find.byType(FlexibleBottomSheet<void>), findsOneWidget);
 
-          await tester.tapAt(const Offset(20.0, 20.0));
+          await tester.tapAt(const Offset(20, 20));
           await tester.pumpAndSettle();
           expect(
-            find.byType(FlexibleBottomSheet),
+            find.byType(FlexibleBottomSheet<void>),
             defaultBoolTestVariant.currentValue!
                 ? findsNothing
                 : findsOneWidget,
@@ -145,19 +152,19 @@ void main() {
           ));
           await tester.pumpAndSettle();
 
-          expect(find.byType(FlexibleBottomSheet), findsOneWidget);
+          expect(find.byType(FlexibleBottomSheet<void>), findsOneWidget);
 
           await tester.drag(
             find.byType(
-              FlexibleBottomSheet,
+              FlexibleBottomSheet<void>,
               skipOffstage: false,
             ),
-            const Offset(0.0, 300.0),
+            const Offset(0, 300),
           );
           await tester.pumpAndSettle();
 
           expect(
-            find.byType(FlexibleBottomSheet),
+            find.byType(FlexibleBottomSheet<void>),
             defaultBoolTestVariant.currentValue!
                 ? findsNothing
                 : findsOneWidget,
@@ -174,17 +181,17 @@ void main() {
           unawaited(showBottomSheet(isCollapsible: false));
           await tester.pumpAndSettle();
 
-          expect(find.byType(FlexibleBottomSheet), findsOneWidget);
+          expect(find.byType(FlexibleBottomSheet<void>), findsOneWidget);
 
           await tester.drag(
-            find.byType(FlexibleBottomSheet),
+            find.byType(FlexibleBottomSheet<void>),
             const Offset(0, -800),
           );
 
           await tester.pumpAndSettle();
 
           await tester.drag(
-            find.byType(FlexibleBottomSheet),
+            find.byType(FlexibleBottomSheet<void>),
             const Offset(0, -800),
           );
 
@@ -193,6 +200,36 @@ void main() {
           final fractionalHeight = getFractionalHeight(tester);
 
           expect(fractionalHeight, moreOrLessEquals(0.8));
+        },
+      );
+
+      testWidgets(
+        'Show SnackBar with Scaffold in BottomSheet tree',
+        (tester) async {
+          await tester.pumpWidget(app);
+
+          showSnackBar();
+
+          unawaited(showBottomSheet());
+
+          await tester.pumpAndSettle();
+
+          expect(find.byType(SnackBar), findsNWidgets(2));
+        },
+      );
+
+      testWidgets(
+        'Show SnackBar without Scaffold in BottomSheet tree',
+        (tester) async {
+          await tester.pumpWidget(app);
+
+          showSnackBar();
+
+          unawaited(showBottomSheet(useRootScaffold: false));
+
+          await tester.pumpAndSettle();
+
+          expect(find.byType(SnackBar), findsOneWidget);
         },
       );
 
@@ -236,13 +273,13 @@ void main() {
 
               await tester.drag(
                 find.byType(
-                  FlexibleBottomSheet,
+                  FlexibleBottomSheet<void>,
                 ),
                 offset,
               );
               await tester.pumpAndSettle();
 
-              expect(find.byType(FlexibleBottomSheet), findsOneWidget);
+              expect(find.byType(FlexibleBottomSheet<void>), findsOneWidget);
 
               final fractionalHeight = getFractionalHeight(tester);
 
@@ -303,8 +340,7 @@ void main() {
 
           await tester.pumpAndSettle();
 
-          final testBinding = tester.binding;
-          testBinding.window.viewInsetsTestValue = const FakeWindowPadding();
+          tester.view.viewInsets = const FakeWindowPadding();
 
           await tester.pumpAndSettle();
 
@@ -324,8 +360,7 @@ void main() {
         (tester) async {
           await tester.pumpWidget(app);
 
-          final testBinding = tester.binding;
-          testBinding.window.viewInsetsTestValue = const FakeWindowPadding();
+          tester.view.viewInsets = const FakeWindowPadding();
 
           unawaited(showBottomSheet(
             anchors: [0.2, 0.5, 0.8],
@@ -475,7 +510,7 @@ final ValueVariant<_AnchorsTestScenario> _anchorsTestVariants =
   },
 );
 
-class FakeWindowPadding implements WindowPadding {
+class FakeWindowPadding implements FakeViewPadding {
   @override
   final double left;
 
@@ -489,9 +524,9 @@ class FakeWindowPadding implements WindowPadding {
   final double bottom;
 
   const FakeWindowPadding({
-    this.left = 0.0,
-    this.top = 0.0,
-    this.right = 0.0,
-    this.bottom = 30.0,
+    this.left = 0,
+    this.top = 0,
+    this.right = 0,
+    this.bottom = 30,
   });
 }